Fix more code quality issues
There are several blocks of code where the happy path was above the guard clause.

Changed it to make the happy path more prominent and the guard clause to be at the top.

Dealing with the unusual conditions as soon as they are detected helps improve readibility for the reader.