Add more error handling
Error handling for file path that can't be found and incorrect
tasks syntax.

Previously when my file path is not found, I system.print in the
terminal and the list would end up empty.

The user might be confused as they wont see the message in the
terminal and would be wondering why the list is empty and not read
from a file. The message that appears with the greeting that the
file can't be found would indicate to them what is missing/wrong in
the code.

Similarly, my chatbot would only print a general message when one
of the todo, deadline, event tasks are inputted but incorrectly
syntaxed/empty. This needed to be warned specificly as well.

This is because the user might not be familiar with the syntax
required for this chatbot and therefore need to be told what they
are doing wrong.